Made arena's size fixed
If the game is going to be multiplayer, it doesn't make sense to have the
arena's size be calculated depending on the camera's aspect ratio (which
would be the host's). However, instead of deleting the class and replacing
all the references in the code base, it is better to keep the class to
have a central place from where to still get the arena's width and height.