Implemented multiplayer for spawned objects
Added network awareness to spawned magnetized objects.
In order for the GameManager to be able to spawn the objects across the
network, its instance needs to be a NetworkIdentity. However, a
NetworkManager object cannot have that component. Consequently, the two
components (GameManager and NetworkManager) have been separated in two
different GameObjects.
Furthermore, with the GameManager being a NetworkIdentity, its state is
handled by the NetworkManager, and so it will be disabled while the game
is not connected. Therefore, there is no need for a paused|playing state,
since the Update function will not be called when the GameManager is
disabled (when there is no active game).
The MagnetizedByPlayer has been expanded for its use in multiplayer
(objects will be magnetized by the closest player).