Implemented vortex mechanic
With long cooldown times, a vortex is opened that drags all objects
towards itself, destroying them when reaching its centre.
This is meant as a high-impact mechanic in the game that breaks the
monotony of the rest of the mechanics, bringing an element of danger to
the game, as players can be destroyed this way too (the only way a player
can die in the game).