Objects inside deposits are not dragged by vortex
'Repelled' objects are not affected by the vortex if inside an active
deposit (if a deposit does not have an assigned player, objects will be
dragged by the vortex even if they appear to be inside).
Added network awareness to corner deposits so the clients don't try
applying the DraggedByVortex force, which is then corrected by the server's
authority, but makes the objects vibrate.