Added GameOver detection upon player death
When a player dies, their client will show the Defeat UI screen, while
remaining in the game scene as an observer.
When only one player stands, they are proclaimed as victors.
Fixed some issues with Server/Client disconnections and the synching of
this to other game instances in the network.