Fixed a bug with ingame score indicator in clients
Fixed a bug that caused the score indicator to be delayed by one unit in
clients. This is because the RPC method is being called before the
SyncVars are transmitted across the network. Instead of this, SyncVarHooks
are now used. However, one must remember to manually assign the values in
the hook methods - otherwise, it won't work.