Balanced the vortex mechanic
Made the vortex appear less times, as it is meant to be a high-impact rare
situation the player has to face, just to break the monotony of the rest
of the game. Before it happened to often, and you had no time to play the
game outside of trying to hide from the vortex.